requirements.txt
selenium
python-vlc
